Move getters into separate source file
