+++ CMakeLists.txt
@@ -27,6 +27,7 @@ add_library(${TARGET} STATIC
     src/debug.h
     src/functions.c
     src/gc_gl.c
+    src/getters.c
     src/image_DXT.c
     src/image_DXT.h
     src/opengx.h
